Enable Date Filter for DataObjects
