+++ src/DataIndex/Filter/DatetimeFilter.php
@@ -11,15 +11,14 @@ declare(strict_types=1);
  *  @license    Pimcore Open Core License (POCL)
  */
 
-namespace Pimcore\Bundle\StudioBackendBundle\DataIndex\Filter\Asset\System;
+namespace Pimcore\Bundle\StudioBackendBundle\DataIndex\Filter;
 
 use Pimcore\Bundle\StudioBackendBundle\DataIndex\Filter\Asset\IsAssetFilterTrait;
-use Pimcore\Bundle\StudioBackendBundle\DataIndex\Filter\FilterInterface;
-use Pimcore\Bundle\StudioBackendBundle\DataIndex\Query\AssetQueryInterface;
 use Pimcore\Bundle\StudioBackendBundle\DataIndex\Query\QueryInterface;
 use Pimcore\Bundle\StudioBackendBundle\Exception\Api\InvalidArgumentException;
 use Pimcore\Bundle\StudioBackendBundle\Grid\Column\ColumnType;
 use Pimcore\Bundle\StudioBackendBundle\MappedParameter\Filter\ColumnFilter;
+use Pimcore\Bundle\StudioBackendBundle\MappedParameter\Filter\ColumnFiltersParameterInterface;
 use function is_array;
 
 /**
@@ -31,21 +30,18 @@ final class DatetimeFilter implements FilterInterface
 
     public function apply(mixed $parameters, QueryInterface $query): QueryInterface
     {
-        $parameters = $this->validateParameterType($parameters);
-        $assetQuery = $this->validateQueryType($query);
-
-        if (!$parameters || !$assetQuery) {
+        if (!$parameters instanceof ColumnFiltersParameterInterface) {
             return $query;
         }
 
         foreach ($parameters->getColumnFilterByType(ColumnType::SYSTEM_DATETIME->value) as $column) {
-            $assetQuery = $this->applyDatetimeFilter($column, $assetQuery);
+            $query = $this->applyDatetimeFilter($column, $query);
         }
 
-        return $assetQuery;
+        return $query;
     }
 
-    private function applyDatetimeFilter(ColumnFilter $column, AssetQueryInterface $query): AssetQueryInterface
+    private function applyDatetimeFilter(ColumnFilter $column, QueryInterface $query): QueryInterface
     {
         if (!is_array($column->getFilterValue())) {
             throw new InvalidArgumentException('Filter value for this filter must be an array');
